generalized SocialMediaShareDialog
